+++ CHANGELOG.md
@@ -58,6 +58,8 @@
 - #1347 Fixing issue when reading orc metadata from DATE dtype
 - #1338 Window Function support for LEAD and LAG statements 
 - #1362 give useful message when file extension is not recognized
+- #1361 Supporting first_value and last_value for Window Function
+
 
 ## Improvements
 - #1293 Added optional acknowledgments to message sending